pyproject.toml
[tool.poetry]
name = "factory_boy-stubs"
version = "0.1.0"
description = ""
authors = ["Alessio Bogon <youtux@gmail.com>"]
packages = [{ include = "factory-stubs", from="src" }, { include="factory_mypy.py", from="src"}]

[tool.poetry.dependencies]
python = "^3.7"
factory-boy = ">=3.2.0"

[tool.poetry.dev-dependencies]
mypy = "^0.961"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
target-version = ['py37', 'py38', 'py39', 'py310']

[tool.isort]
profile = 'black'

[tool.mypy]
allow_redefinition = false
check_untyped_defs = true
disallow_untyped_decorators = true
disallow_any_explicit = false
disallow_any_generics = true
disallow_untyped_calls = true
disallow_untyped_defs = true
ignore_errors = false
ignore_missing_imports = true
implicit_reexport = false
strict_optional = true
no_implicit_optional = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unreachable = true
warn_no_return = true
warn_return_any = true
pretty = true
show_error_codes = true
